Fit a GP dynamic factor model

Fit a squared exponential GP dynamic factor model using Stan

Usage

GPDF(
  formula,
  family = gaussian(),
  h = 1,
  chains = 4,
  cores = 1,
  iter = 1000,
  warmup = floor(iter/2),
  adapt_delta = 0.8,
  max_treedepth = 10,
  ...
)

Arguments

formula

Model specification (see "Specials" section)

family

A family object specifying the outcome distribution to use in fitting. Currently only gaussian() and scat() (i.e. Student-T) are supported

h

integer specifying the forecast horizon

chains

integer specifying the number of chains to be run

cores

integer specifying the number of parallel cores to use

iter

integer specifying the total number of iterations to run per chain (including warmup)

warmup

integer specifying the number of initial iterations to use as burnin

adapt_delta

the thin of the jumps in a HMC method

max_treedepth

maximum tree depth per iteration

...

other arguments to pass to rstan::sampling()

Value

A model specification

Author

Nicholas J Clark

Examples

if (FALSE) { # \dontrun{
# Fit a functional forecasting model, then use GPDF for forecasting
library(dplyr)

# Split growth data into training and test sets
train_data <- growth_data %>% filter(age_yr <= 16)
test_data <- growth_data %>% filter(age_yr > 16)

# Step 1: Fit ffc_gam model with time-varying coefficients
mod <- ffc_gam(
  height_cm ~ fts(age_yr, by = id, time_k = 5),
  data = train_data,
  time = "age_yr",
  family = gaussian()
)

# Step 2: Use GPDF for forecasting functional coefficients
fc <- forecast(mod, newdata = test_data, model = "GPDF", 
               chains = 1, iter = 500)
} # }