Fit a functional time series model using dynamic functional coefficients

Fit Generalized Additive Models that can include time-varying (dynamic) functions

ffc_gam(
  formula,
  family = gaussian(),
  data = list(),
  time,
  engine = c("gam", "bam"),
  ...
)

Arguments

formula

A GAM formula (see formula.gam and also gam.models). This is exactly like the formula for a GLM except that smooth terms, fts(), s() and te() can be added to the right hand side to specify that the linear predictor depends on smooth functions of predictors (or linear functionals of these).

family

This is a family object specifying the distribution and link to use in fitting etc. See glm and family for more details. The extended families listed in family.mgcv can also be used.

data

A data frame or list containing the model response variable and covariates required by the formula. By default the variables are taken from environment(formula): typically the environment from which gam is called.

time

character specifying which variable in data represents the the time ordering of the observations

engine

character string specifying which mgcv interface to use for fitting the model.

...

other arguments to pass to either gam

Value

An object of class ffc_gam, which inherits from objects of class gam or bam

Details

This function will update the supplied formula to ensure any time-varying functionals (supplied through fts() terms in the formula right hand side) are appropriately incorporated into the model. It then passes the updated model and data objects to the specified engine for model fitting

See also

Author

Nicholas J Clark

Examples

# Fit a dynamic function-on-function regression to the Queensland
# mortality data
data("qld_mortality")
mod <- ffc_gam(
  deaths ~
  offset(log(population)) +
  sex +
  fts(age, k = 8, bs = 'cr',
     time_bs = 'cr', time_k = 10),
 time = 'year',
 data = qld_mortality,
 family = poisson(),
 engine = 'bam'
)
class(mod)
#> [1] "ffc_gam" "bam"     "gam"     "glm"     "lm"     
summary(mod)
#> 
#> Family: poisson 
#> Link function: log 
#> 
#> Formula:
#> deaths ~ sex + offset(log(population)) + s(year, by = fts_bs_s_age__1, 
#>     bs = "cr", k = 10, id = 1) + s(year, by = fts_bs_s_age__2, 
#>     bs = "cr", k = 10, id = 1) + s(year, by = fts_bs_s_age__3, 
#>     bs = "cr", k = 10, id = 1) + s(year, by = fts_bs_s_age__4, 
#>     bs = "cr", k = 10, id = 1) + s(year, by = fts_bs_s_age__5, 
#>     bs = "cr", k = 10, id = 1) + s(year, by = fts_bs_s_age__6, 
#>     bs = "cr", k = 10, id = 1) + s(year, by = fts_bs_s_age__7, 
#>     bs = "cr", k = 10, id = 1) + s(year, by = fts_bs_fts_age1_mean, 
#>     bs = "cr", k = 10, id = 1)
#> 
#> Parametric coefficients:
#>              Estimate Std. Error z value Pr(>|z|)    
#> (Intercept) -5.520317   0.002436 -2266.2   <2e-16 ***
#> sexmale      0.474023   0.002078   228.1   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Approximate significance of smooth terms:
#>                                edf Ref.df Chi.sq p-value    
#> s(year):fts_bs_s_age__1      5.046  5.955  59933  <2e-16 ***
#> s(year):fts_bs_s_age__2      6.123  7.197  48275  <2e-16 ***
#> s(year):fts_bs_s_age__3      6.578  7.678  26531  <2e-16 ***
#> s(year):fts_bs_s_age__4      7.422  8.476   8841  <2e-16 ***
#> s(year):fts_bs_s_age__5      7.939  8.894 191290  <2e-16 ***
#> s(year):fts_bs_s_age__6      7.861  8.791 629392  <2e-16 ***
#> s(year):fts_bs_s_age__7      6.219  7.290 632787  <2e-16 ***
#> s(year):fts_bs_fts_age1_mean 8.226  8.707  12624  <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> R-sq.(adj) =  0.975   Deviance explained = 95.9%
#> fREML =  33227  Scale est. = 1         n = 8282

# Predictions work in the usual way
head(predict(mod, type = 'link'))
#>         1         2         3         4         5         6 
#> -5.856545 -6.080781 -6.303054 -6.521396 -6.733846 -6.938437 
head(predict(mod, type = 'response'))
#>            1            2            3            4            5            6 
#> 0.0028611125 0.0022863893 0.0018307061 0.0014716126 0.0011899479 0.0009697844